SourceCode/requirements.txt
# Biomapas dependencies

b-cfn-lambda-layer==1.1.3

# 3rd party other dependencies

pynamodb==5.2.1

# AWS CDK dependencies

aws-cdk.core==1.140.0
aws-cdk.assets==1.140.0
aws-cdk.assertions==1.140.0
aws-cdk.aws-lambda==1.140.0
aws-cdk.aws-dynamodb==1.140.0
